requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
certifi==2026.2.25
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.7
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
distro==1.9.0
feedparser==6.0.12
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-pasta==0.2.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
identify==2.6.15
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
moviepy==2.2.1
namex==0.1.0
nodeenv==1.9.1
numpy==2.4.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pre_commit==4.3.0
proglog==0.1.12
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uv==0.8.18
virtualenv==20.35.3
websockets==16.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
yt-dlp==2026.1.31
